requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
psycopg2-binary==2.9.5
transformers==4.37.2
torch==2.2.0
accelerate==0.27.2
bitsandbytes==0.42.0
sentencepiece==0.1.99
protobuf==4.25.2
PyJWT==2.8.0 